monkey en sea
Monkeys are not typically found in the sea, as they are not adapted to swimming or living underwater. However, there are some species of primates, such as macaques and baboons, that are known to be good swimmers and may venture into water sources such as rivers and lakes. Additionally, some primates, such as the proboscis monkey, live near coastal areas and may swim short distances in rivers or in the sea to find food. But in general, monkeys are not aquatic animals and are not found in the sea.